refactor: package all model requester providers
